requirements.txt
pandas
python-dotenv
python-telegram-bot==21.2
streamlit
openpyxl
